requirements.txt
ctkcomponents==0.4
CTkMessagebox==2.5
customtkinter==5.2.2
loguru==0.7.2
Pillow==10.3.0
pymongo==4.7.0
bcrypt==4.1.3
pywinstyles==1.8
CTkColorPicker==0.9.0